+++ spezspellz/models/__init__.py
@@ -1,0 +1,37 @@
+"""This file contains models."""
+from .category import Category
+from .user_info import UserInfo
+from .spell import Spell
+from .tag import Tag
+from .has_tag import HasTag
+from .rate import Rate
+from .rate_tag import RateTag
+from .rate_category import RateCategory
+from .spell_notification import SpellNotification
+from .attachment import Attachment
+from .bookmark import Bookmark
+from .review import Review
+from .spell_history_entry import SpellHistoryEntry
+from .review_comment import ReviewComment
+from .spell_comment import SpellComment
+from .comment_comment import CommentComment
+
+
+__all__ = (
+    "Category",
+    "UserInfo",
+    "Spell",
+    "Tag",
+    "HasTag",
+    "Rate",
+    "RateTag",
+    "RateCategory",
+    "SpellNotification",
+    "Attachment",
+    "Bookmark",
+    "Review",
+    "SpellHistoryEntry",
+    "ReviewComment",
+    "SpellComment",
+    "CommentComment"
+)
